S3 us7 securisation des routes

+++ src/Controller/SkillController.php
@@ -6,25 +6,42 @@ use App\Entity\Skill;
 use App\Form\SkillType;
 use App\Repository\SkillRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route('/skill')]
+#[Route('/skill', name: 'skill_')]
 class SkillController extends AbstractController
 {
-    #[Route('/', name: 'app_skill_index', methods: ['GET'])]
+    private Security $security;
+
+    public function __construct(
+        Security $security
+    ) {
+        $this->security = $security;
+    }
+
+    #[Route('/', name: 'index', methods: ['GET'])]
     public function index(SkillRepository $skillRepository): Response
     {
+        if (!($this->security->isGranted('ROLE_COLLABORATEUR'))) {
+            return $this->redirectToRoute('app_home');
+        }
+
         return $this->render('skill/index.html.twig', [
             'skills' => $skillRepository->findAll(),
         ]);
     }
 
-    #[Route('/new', name: 'app_skill_new', methods: ['GET', 'POST'])]
+    #[Route('/new', name: 'new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!($this->security->isGranted('ROLE_COLLABORATEUR'))) {
+            return $this->redirectToRoute('app_home');
+        }
+
         $skill = new Skill();
         $form = $this->createForm(SkillType::class, $skill);
         $form->handleRequest($request);
@@ -33,7 +50,7 @@ class SkillController extends AbstractController
             $entityManager->persist($skill);
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_skill_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('skill_index', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->render('skill/new.html.twig', [
@@ -42,24 +59,32 @@ class SkillController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_skill_show', methods: ['GET'])]
+    #[Route('/{id}', name: 'show', methods: ['GET'])]
     public function show(Skill $skill): Response
     {
+        if (!($this->security->isGranted('ROLE_COLLABORATEUR'))) {
+            return $this->redirectToRoute('app_home');
+        }
+
         return $this->render('skill/show.html.twig', [
             'skill' => $skill,
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_skill_edit', methods: ['GET', 'POST'])]
+    #[Route('/{id}/edit', name: 'edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Skill $skill, EntityManagerInterface $entityManager): Response
     {
+        if (!($this->security->isGranted('ROLE_COLLABORATEUR'))) {
+            return $this->redirectToRoute('app_home');
+        }
+
         $form = $this->createForm(SkillType::class, $skill);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_skill_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('skill_index', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->render('skill/edit.html.twig', [
@@ -68,14 +93,18 @@ class SkillController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_skill_delete', methods: ['POST'])]
+    #[Route('/{id}', name: 'delete', methods: ['POST'])]
     public function delete(Request $request, Skill $skill, EntityManagerInterface $entityManager): Response
     {
+        if (!($this->security->isGranted('ROLE_COLLABORATEUR'))) {
+            return $this->redirectToRoute('app_home');
+        }
+
         if ($this->isCsrfTokenValid('delete' . $skill->getId(), $request->request->get('_token'))) {
             $entityManager->remove($skill);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('app_skill_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('skill_index', [], Response::HTTP_SEE_OTHER);
     }
 }
